+++ R/trialmaster.R
@@ -22,9 +22,9 @@ read_trialmaster = function(archive, ..., use_cache=FALSE,
                             clean_names_fun=NULL,
                             split_mixed=FALSE,
                             extend_lookup=TRUE,
-                            key_columns=get_key_cols(),
                             pw=getOption("trialmaster_pw"), 
-                            verbose=getOption("edc_verbose", 1)){
+                            verbose=getOption("edc_verbose", 1),
+                            key_columns="deprecated"){
   directory = dirname(archive)
   extract_datetime = parse_file_datetime(archive)
   check_dots_empty()
@@ -94,8 +94,8 @@ read_trialmaster = function(archive, ..., use_cache=FALSE,
 #' @param ... unused
 #' @param split_mixed \[`logical(1): FALSE`]\cr whether to split mixed datasets. See [split_mixed_datasets]. 
 #' @param extend_lookup \[`character(1): FALSE`]\cr whether to enrich the lookup table. See [extend_lookup].
-#' @param key_columns \[`list`]\cr the result of [get_key_cols()], containing the column name used for patient ID and CRF name. Important for `split_mixed` and `extend_lookup`.
 #' @param clean_names_fun \[`function`]\cr a function to clean column names, e.g. [janitor::clean_names()]
+#' @param key_columns deprecated
 #'
 #' @return a list containing one dataframe for each `.xpt` file in the folder, the extraction date (`datetime_extraction`), and a summary of all imported tables (`.lookup`). If not set yet, option `edc_lookup` is automatically set to `.lookup`.
 #' @export
@@ -103,7 +103,7 @@ read_trialmaster = function(archive, ..., use_cache=FALSE,
 #' @importFrom dplyr across mutate na_if
 #' @importFrom forcats as_factor
 #' @importFrom haven read_xpt
-#' @importFrom purrr imap keep map_lgl
+#' @importFrom purrr imap keep keep_at map_lgl
 #' @importFrom rlang check_dots_empty is_error set_names
 #' @importFrom stringr str_remove
 #' @importFrom tibble as_tibble
@@ -112,24 +112,23 @@ read_tm_all_xpt = function(directory, ..., format_file="procformat.sas",
                            clean_names_fun=NULL, 
                            split_mixed=FALSE,
                            extend_lookup=TRUE,
-                           key_columns=get_key_cols(),
                            datetime_extraction=NULL, 
-                           verbose=getOption("edc_verbose", 1)){
+                           verbose=getOption("edc_verbose", 1),
+                           key_columns="deprecated"){
   check_dots_empty()
   reset_manual_correction()
   clean_names_fun = get_clean_names_fun(clean_names_fun)
-  patient_id = key_columns$patient_id
-  stopifnot(is.character(patient_id))
   
   datasets = dir(directory, pattern = "\\.xpt$", full.names=TRUE)
   datasets_names = basename(datasets) %>% str_remove("\\.xpt")
   if(is.null(datetime_extraction)) datetime_extraction=get_folder_datetime(directory)
   
+  #reading
   rtn = datasets %>% 
     set_names(tolower(datasets_names)) %>% 
     imap(~tryCatch(read_xpt(.x), error=function(e) e))
-  id_found = map_lgl(rtn, ~any(tolower(patient_id) %in% tolower(names(.x))))
   
+  #applying formats
   if(!is.null(format_file)){
     procformat = file.path2(directory, format_file)
     if(!file.exists(procformat)) procformat = format_file
@@ -150,6 +149,7 @@ read_tm_all_xpt = function(directory, ..., format_file="procformat.sas",
       })
   }
   
+  #strip out non-UTF8 (and warn if needed)
   .lookup = get_lookup(rtn)
   bad_utf8 = check_invalid_utf8(.lookup, warn=verbose>0)
   bad_utf8 %>%
@@ -159,12 +159,31 @@ read_tm_all_xpt = function(directory, ..., format_file="procformat.sas",
     })
   .lookup = get_lookup(rtn)
   
-  if(isTRUE(split_mixed)){
-    if(!any(id_found)) cli_warn("Patient ID column {.val {patient_id}} was not found in any dataset. Is it possible that you made a spelling mistake?")
-    mixed = split_mixed_datasets(patient_id, datasets=rtn, verbose=FALSE)
-    rtn = c(rtn, mixed)
+  #split mixed datasets (with short and long format)
+  key_columns = get_key_cols(.lookup)
+  patient_id = key_columns$patient_id
+  id_found = map_lgl(rtn, ~any(tolower(patient_id) %in% tolower(names(.x))))
+  if(!isFALSE(split_mixed)){
+    split_mixed_names = split_mixed
+    if(isTRUE(split_mixed)) split_mixed_names = names(rtn)
+    if(!is.character(split_mixed)){
+      cli_abort("{.arg split_mixed} should be either FALSE, TRUE, or a character vector of column names.")
+    }
+    if(any(id_found)){
+      mixed = rtn %>% 
+        keep_at(split_mixed_names) %>% 
+        split_mixed_datasets(id=patient_id, verbose=FALSE)
+      if(!isTRUE(split_mixed) && length(mixed)==0){
+        cli_warn("Dataset{?s} {.val {split_mixed_names}} are not mixed (either short or long) and cannot be split", 
+                 class="edc_read_cannot_split_mixed_warn")
+      }
+      rtn = c(rtn, mixed)
+    } else {
+      cli_warn("Patient ID column {.val {patient_id}} was not found in any dataset")
+    }
   }
   
+  #manage errors
   errs = keep(rtn, is_error)
   if(length(errs)>0){
     cli_warn(c("SAS dataset{?s} {.val {names(errs)}} could not be read from 
@@ -179,10 +198,7 @@ read_tm_all_xpt = function(directory, ..., format_file="procformat.sas",
   rtn$.lookup = .lookup
     
   if(isTRUE(extend_lookup)){
-    if(!any(id_found)) cli_warn("Patient ID column {.val {patient_id}} was not found in any dataset. Is it possible that you made a spelling mistake?")
-    rtn$.lookup = extend_lookup(rtn$.lookup, datasets=rtn, 
-                                key_columns=get_key_cols(patient_id=patient_id, 
-                                                         crfname=key_columns$crfname))
+    rtn$.lookup = extend_lookup(rtn$.lookup, datasets=rtn)
   }
   
   if(!is.null(getOption("edc_lookup", NULL))){
@@ -195,8 +211,6 @@ read_tm_all_xpt = function(directory, ..., format_file="procformat.sas",
 
 
 # Utils ---------------------------------------------------------------------------------------
-
-
 
 #' @noRd
 #' @keywords internal